backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
fpdf
jinja2
python-multipart
python-dotenv
google-genai
pytest
playwright
trafilatura
beautifulsoup4
feedparser
readability-lxml
python-jose[cryptography]
passlib[bcrypt]
bcrypt
notion-client
httpx
lxml
requests
lxml_html_clean
celery
redis
pydantic
pydantic-settings
pydantic_core
greenlet
grpcio-status
pyee
itemadapter
certifi
anyio
starlette
wheel
setuptools
pip
markdown
croniter
APScheduler
